Added base files for other pages
The whole "copypasting the nav bar" instead of including it by other
means is a little annoying. In this case I think it's worth it though since
it'll be served as static pages from github - having a backend to
include stuff isn't possible, and including more stuff through jquery is
not attractive to me since I'm trying to keep this page as lean as
possible.